fix: restore u64 casts for macOS cross-platform compatibility
On macOS, f_blocks and f_bavail are u32 while f_frsize is u64.
The cast is necessary for cross-platform compatibility.
Added #[allow(clippy::unnecessary_cast)] with comment explaining why.

+++ src/disk_space.rs
@@ -93,9 +93,12 @@ pub fn get_disk_space(path: &Path) -> Result<DiskSpaceInfo> {
     }
 
     // Calculate values from statvfs
-    let frsize = stat.f_frsize;
-    let total_bytes = stat.f_blocks * frsize;
-    let available_bytes = stat.f_bavail * frsize;
+    #[allow(clippy::unnecessary_cast)] // Cast needed for macOS (u32) vs Linux (u64)
+    let frsize = stat.f_frsize as u64;
+    #[allow(clippy::unnecessary_cast)] // Cast needed for macOS (u32) vs Linux (u64)
+    let total_bytes = stat.f_blocks as u64 * frsize;
+    #[allow(clippy::unnecessary_cast)] // Cast needed for macOS (u32) vs Linux (u64)
+    let available_bytes = stat.f_bavail as u64 * frsize;
     let used_bytes = total_bytes - available_bytes;
     let used_percent = if total_bytes > 0 {
         (used_bytes as f64 / total_bytes as f64) * 100.0
